+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/control/DefaultCommandControlTest.java
@@ -63,7 +63,7 @@ public final class DefaultCommandControlTest {
 						.background(Color.BLACK)
 						.font(font)
 						.build();
-		JButton button = button(control).build();
+		JButton button = button().control(control).build();
 		assertFalse(button.isEnabled());
 		SwingUtilities.invokeLater(() -> {
 			enabledState.set(true);