+++ changelog.md
@@ -89,6 +89,7 @@ Codion Change Log
 - TemporalField.Builder.TemporalClassBuilder added.
 - TemporalFieldPanel.Builder.TemporalClassBuilder added.
 - NumberField.Builder.NumberClassBuilder added.
+- ButtonBuilder, overloaded builder() methods removed.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui